Decrease severity for illegal plugin ID to warning
Existing plugins do not have a reasonable option to fix such issues. On the other hand, new plugins will get a warning and a opportunity to fix this.
