+++ intellij-plugin-structure/structure-intellij/src/main/java/com/jetbrains/plugin/structure/intellij/problems/UnacceptableWarnings.kt
@@ -22,17 +22,3 @@ class HttpLinkInDescription(private val link: String) : InvalidDescriptorProblem
     get() = "All links in description must be HTTPS: $link"
 }
 
-open class IllegalPluginId(private val illegalPluginId: String) : InvalidDescriptorProblem("id") {
-
-  override val level
-    get() = Level.UNACCEPTABLE_WARNING
-
-  override val detailedMessage
-    get() = "Plugin ID '$illegalPluginId' is not valid. See https://plugins.jetbrains.com/docs/intellij/plugin-configuration-file.html#idea-plugin__id"
-}
-
-class IllegalPluginIdPrefix(private val illegalPluginId: String, private val illegalPrefix: String) : IllegalPluginId(illegalPluginId) {
-  override val detailedMessage
-    get() = "Plugin ID '$illegalPluginId' has an illegal prefix '$illegalPrefix'. See https://plugins.jetbrains.com/docs/intellij/plugin-configuration-file.html#idea-plugin__id"
-}
-
